Moved everything from `index.rs` and `manufacture.rs` into `main.rs`, `logic.rs` and `input.rs`. Awaiting next step
