+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod renderer;
+pub mod logic;
+pub mod input;
 pub mod system;
-pub mod manufacture;
 pub mod world;
 pub mod player;
